# other/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Welcome to the nodejs backend of Booboky!"
  version: "1.0.0"
  title: "Booboky"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /author/{id}/review:
    get:
      tags:
      - "author"
      - "review"
      summary: "Get reviews for a given author"
      operationId: "authorReviewGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Bad author id"
        404:
          description: "Status 404"
      x-swagger-router-controller: "Author"
    post:
      tags:
      - "book"
      - "review"
      summary: "Add new review for the given author"
      operationId: "authorReviewPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "userID"
        in: "query"
        description: "The user that wrote the review"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        200:
          description: "Status 200"
      security:
      - OAuth2:
        - "user"
        - "admin"
      x-swagger-router-controller: "Book"
    delete:
      tags:
      - "author"
      - "review"
      summary: "Delete a review from a given user"
      operationId: "authorReviewDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "reviewID"
        in: "query"
        description: "The id of the review to be removed"
        required: true
        type: "integer"
      responses:
        200:
          description: "Status 200"
        400:
          description: "invalid review id"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "Author"
...
  /book/{id}/review:
    get:
      tags:
      - "author"
      - "review"
      summary: "Get reviews for a given book"
      operationId: "bookReviewGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
        400:
          description: "Bad author id"
        404:
          description: "Status 404"
      x-swagger-router-controller: "Author"
    post:
      tags:
      - "book"
      - "review"
      summary: "Add new review for the given book"
      operationId: "bookReviewPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "userID"
        in: "query"
        description: "The user that wrote the review"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        200:
          description: "Status 200"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "Book"
    delete:
      tags:
      - "author"
      - "review"
      summary: "Delete a review from a given user"
      operationId: "bookReviewDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "reviewID"
        in: "query"
        description: "The id of the review to be removed"
        required: true
        type: "integer"
      responses:
        200:
          description: "Status 200"
        400:
          description: "invalid review id"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "Author"
...
  /user/{id}/chart:
    get:
      tags:
      - "user"
      - "chart"
      summary: "Get the user's current chart"
      operationId: "userChartGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            $ref: "#/definitions/Chart"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      - "chart"
      summary: "Updates the current chart"
      operationId: "userChartPUT"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_1"
      responses:
        200:
          description: "Status 200"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      - "chart"
      summary: "Delete the current chart"
      operationId: "userChartDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
  /event:
    get:
      tags:
      - "event"
      summary: "Get the complete list of events"
      operationId: "eventGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        404:
          description: "Status 404"
      x-swagger-router-controller: "Event"
    post:
      tags:
      - "event"
      summary: "Add new event"
      operationId: "eventPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        200:
          description: "Status 200"
        400:
          description: "invalid event object"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
      x-swagger-router-controller: "Event"
  /event/{id}:
    get:
      tags:
      - "event"
      summary: "Get event for a given id"
      operationId: "eventIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            $ref: "#/definitions/Event"
        404:
          description: "Status 404"
      x-swagger-router-controller: "Event"
    put:
      tags:
      - "event"
      summary: "Update an event"
      operationId: "eventIdPUT"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        200:
          description: "Status 200"
        400:
          description: "Status 400"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
      x-swagger-router-controller: "Event"
    delete:
      summary: "Delete an event"
      operationId: "eventIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
      x-swagger-router-controller: "Default"
  /event/search:
    get:
      tags:
      - "event"
      summary: "Search for an event"
      operationId: "eventSearchGET"
      produces:
      - "application/json"
      parameters:
      - name: "query_string"
        in: "query"
        description: "The search query"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "The name of the event"
        required: false
        type: "string"
      - name: "author_name"
        in: "query"
        description: "The name of the author for the event"
        required: false
        type: "string"
      - name: "author_id"
        in: "query"
        description: "The id of the author for the event"
        required: false
        type: "integer"
      - name: "book_name"
        in: "query"
        required: false
        type: "string"
      - name: "book_id"
        in: "query"
        required: false
        type: "integer"
      - name: "date"
        in: "query"
        required: false
        type: "string"
        format: "date"
      - name: "date_from"
        in: "query"
        required: false
        type: "string"
        format: "date"
      - name: "date_to"
        in: "query"
        required: false
        type: "string"
        format: "date"
      - name: "Location"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        400:
          description: "At least one query parameter must exist"
        404:
          description: "Status 404"
      x-swagger-router-controller: "Event"
securityDefinitions:
  OAuth2:
    type: "oauth2"
    tokenUrl: "https://localhost/api/auth"
    flow: "password"
    scopes:
      admin: "Website administrator"
      user: "Registred user"
definitions:
  Book:
    type: "object"
    required:
    - "author"
    - "price"
    - "publisher"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      author:
        type: "string"
      price:
        $ref: "#/definitions/Amount"
      image_url:
        type: "array"
        items:
          type: "string"
      genres:
        type: "array"
        items:
          type: "string"
      publisher:
        type: "string"
    description: "A book for sale in the store"
    example: "{\n  id: 0,\n  title: \"Il deserto dei tartari\",\n  author: \"Dino\
      \ Buzzati\",\n  price: {\n    value: 10,\n    currency: \"eur\"\n  },\n  image_url:\
      \ \"https://test.com\"\n}"
  Amount:
    type: "object"
    required:
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000
    description: "Price\n"
    example: "{\n  currency: \"eur\",\n  value: 11\n}"
  User:
    type: "object"
    required:
    - "email"
    - "name"
    - "surname"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
      birthdate:
        type: "string"
        format: "date"
    description: "A user"
    example: "{\n  id: 1,\n  name: \"Alan\",\n  surname: \"Turing\",\n  email: \"\
      alan.turing@example.com\"\n  birthdate: \"1989-01-01\"\n}"
  Review:
    type: "object"
    required:
    - "author"
    - "book_author"
    - "rating"
    - "timestamp"
    - "title"
    properties:
      id:
        type: "integer"
      author:
        type: "integer"
        description: "ID of the user to which this review refers"
      timestamp:
        type: "string"
        format: "date-time"
      title:
        type: "string"
      body:
        type: "string"
      rating:
        type: "number"
      book_author:
        type: "integer"
        description: "ID of the book (or author) to which this review refers."
    description: "This is a review object."
    example:
      author: 6
      rating: 1.46581298050294517310021547018550336360931396484375
      book_author: 5
      id: 0
      title: "title"
      body: "body"
      timestamp: "2000-01-23T04:56:07.000+00:00"
...
  Address:
    type: "object"
    required:
    - "address_line_1"
    - "cap"
    - "city"
    - "country"
    - "name"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      address_line_1:
        type: "string"
      address_line_2:
        type: "string"
      cap:
        type: "integer"
      city:
        type: "string"
      country:
        type: "string"
    description: "An address -.-"
    example:
      country: "country"
      cap: 6
      city: "city"
      name: "name"
      address_line_1: "address_line_1"
      id: 0
      address_line_2: "address_line_2"
  Event:
    type: "object"
    required:
    - "description"
    - "location"
    - "name"
    - "timestamp"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
      location:
        $ref: "#/definitions/Address"
      timestamp:
        type: "string"
        format: "date-time"
      image_urls:
        type: "array"
        description: "Array with the images of the event"
        items:
          type: "string"
    description: "An event, oc -.-"
    example:
      name: "name"
      description: "description"
      image_urls:
      - "image_urls"
      - "image_urls"
      location:
        country: "country"
        cap: 6
        city: "city"
        name: "name"
        address_line_1: "address_line_1"
        id: 0
        address_line_2: "address_line_2"
      id: 0
      timestamp: "2000-01-23T04:56:07.000+00:00"
...
  Chart:
    type: "object"
    required:
    - "Books"
    - "User"
    - "timestamp"
    - "total_amount"
    properties:
      User:
        $ref: "#/definitions/User"
      Books:
        type: "array"
        description: "The list of books ordered"
        items:
          $ref: "#/definitions/Order_Books"
      total_amount:
        type: "number"
        description: "Amount to pay for this order"
      timestamp:
        type: "string"
        format: "date-time"
    description: "An order for some books"
    example: "{\n  OrderID: 123456789,\n  User: \n  {\n   id: 1,\n   name: \"Alan\"\
      ,\n   surname: \"Turing\",\n   email: \"alan.turing@example.com\"\n   birthdate:\
      \ \"1989-01-01\"\n}\n  Books: [\n  {\n  id: 0,\n  title: \"Il deserto dei tartari\"\
      ,\n  author: \"Dino Buzzati\",\n  price: {\n    value: 10,\n    currency: \"\
      eur\"\n  }\n\n},\n{\n  id: 1,\n  title: \"Il deserto dei tartari\",\n  author:\
      \ \"Dino Buzzati\",\n  price: {\n    value: 10,\n    currency: \"eur\"\n  }\n\
      }\n  ]\n}"
...
  body_1:
    type: "object"
    required:
    - "bookID"
    - "qty"
    properties:
      bookID:
        type: "integer"
      qty:
        type: "integer"
  Order_Books:
    required:
    - "book"
    - "qty"
    properties:
      book:
        $ref: "#/definitions/Book"
      qty:
        type: "integer"